portability

+++ client.h
@@ -4,7 +4,7 @@
 #include <vector>
 #include <map>
 
-#include <net/net.h>
+#include <net/poll.h>
 #include <lv2/sysfs.h>
 
 class Client;
@@ -17,36 +17,37 @@ typedef int (*func)(Client*);
 
 class Client {
 private:
-    func data_handler;
-    vector<pollfd>* pollfds;
-    map<int, Client*>* clients;
-    map<int, Client*>* clients_data;
+	func data_handler;
+	vector<pollfd>* pollfds;
+	map<int, Client*>* clients;
+	map<int, Client*>* clients_data;
 
 public:
-    int socket_ctrl;
-    int socket_data;
-    int socket_pasv;
-    char* buffer;
-    char* buffer_data;
-    string lastcmd;
+	int socket_ctrl;
+	int socket_data;
+	int socket_pasv;
+	char* buffer;
+	char* buffer_data;
+	string lastcmd;
 
-    bool cvar_auth;
-    string cvar_user;
-    vector<string> cvar_cwd;
-    string cvar_rnfr;
-    u64 cvar_rest;
-    s32 cvar_fd;
-    bool cvar_use_aio;
-    sysFSAio cvar_aio;
-    s32 cvar_aio_id;
+	bool cvar_auth;
+	string cvar_user;
+	vector<string> cvar_cwd;
+	string cvar_rnfr;
+	size_t cvar_rest;
+	long cvar_fd;
+	bool cvar_use_aio;
+	sysFSAio cvar_aio;
+	long cvar_aio_id;
 
-    Client(int, vector<pollfd>*, map<int, Client*>*, map<int, Client*>*);
-    ~Client(void);
-    void send_string(string);
-    void send_code(int, string);
-    void send_multicode(int, string);
-    void handle_command(map<string, cmdfunc>*, string, string);
-    void handle_data(void);
-    int data_start(func, short events);
-    void data_end(void);
+	Client(int, vector<pollfd>*, map<int, Client*>*, map<int, Client*>*);
+	~Client(void);
+	
+	void send_string(string);
+	void send_code(int, string);
+	void send_multicode(int, string);
+	void handle_command(map<string, cmdfunc>*, string, string);
+	void handle_data(void);
+	int data_start(func, short events);
+	void data_end(void);
 };
